+++ tests/testthat/test-utilities-data-combined.R
@@ -32,18 +32,15 @@ test_that(
   }
 )
 
-test_that("calculateResiduals returns a dataframe", {
-  expect_true(
-    "data.frame" %in% class(calculateResiduals(myDC, scaling = "lin"))
-  )
+test_that("calculateResiduals returns a data frame", {
+  expect_s3_class(calculateResiduals(myDC, scaling = "lin"), "data.frame")
 })
-
-
 
 test_that("calculateResiduals returns expected columns", {
   expected_column_names <- c(
     "group",
     "name",
+    "nameSimulated",
     "xValues",
     "xUnit",
     "xDimension",
@@ -120,3 +117,19 @@ test_that(
     )
   }
 )
+
+test_that("calculateResiduals computes all observed, simulated data pairs correctly", {
+  myDC$addSimulationResults(simData, groups = "myGroup", names = "secondSimResults")
+  residuals <- calculateResiduals(myDC, scaling = "lin")
+
+  expect_setequal(residuals$nameSimulated, c(
+    "Organism|PeripheralVenousBlood|Aciclovir|Plasma (Peripheral Venous Blood)",
+    "secondSimResults"
+  ))
+
+  expect_equal(
+    residuals$residualValues,
+    rep(c(0, 0.1000000, -0.0999999, 0.0, -0.1999998, -1), 2),
+    tolerance = 1e-5
+  )
+})